services/ai-service/requirements.txt
fastapi
uvicorn
pydantic
transformers
torch
python-dotenv
openai
accelerate  
bitsandbytes
groq